Infineon: Switch to 1.9.0 code base, add xmc7000 family support, refactor memory layer

+++ boot/mbed/include/flash_map_backend/flash_map_backend.h
@@ -143,7 +143,7 @@ int flash_area_erase(const struct flash_area * fap, uint32_t off, uint32_t len);
 /*
  * Alignment restriction for flash writes.
  */
-uint8_t flash_area_align(const struct flash_area * fap);
+uint32_t flash_area_align(const struct flash_area * fap);
 
 /*
  * What is value is read from erased flash bytes.
